requirements.txt
alembic==1.13.1
aniso8601==9.0.1
attrs==23.2.0
black==24.3.0
blinker==1.7.0
cachelib==0.12.0
click==8.1.3
Flask==3.0.2
Flask-Cors==4.0.0
flask-restx==1.3.0
Flask-Session==0.7.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==20.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Mako==1.3.2
MarkupSafe==2.1.2
msgspec==0.18.6
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
psycopg2-binary==2.9.5
pytest==8.1.1
python-dotenv==0.19.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.0
rpds-py==0.18.0
six==1.16.0
SQLAlchemy==2.0.27
type_enforced==1.3.0
typing_extensions==4.10.0
Werkzeug==3.0.2